Added new loops to the test harness for complex functions. These generate the tables on page 120 for the Q(l,m) independently from the P(l,m).
Fixed the computation of the factor zcq0 in regard to (z + 1) / (z - 1)
as explained in the update to the previous update which was to the
zhang_and_jin.cxx non template version of the code.

Ok, so the computed Q(l,m) are now real and onlyo real for z=(2,0) as they
should be.

Noting still have errors in the actual real values. Must be the loop
structure as per second bug fix in the previous update.